fix confirm reservasi

+++ app/Http/Controllers/Jadwal/JadwalController.php
@@ -16,10 +16,10 @@ class JadwalController extends Controller
      */
     public function index()
     // {   if (Auth::user()->role == 1)
-        {
-            $jadwal = Jadwal::get();
-            return JadwalResource::collection($jadwal);
-            return $jadwal;
+    {
+        $jadwal = Jadwal::get();
+        return JadwalResource::collection($jadwal);
+        return $jadwal;
         // }
         // return['Anda tidak memiliki akses'];
     }
@@ -28,41 +28,40 @@ class JadwalController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {   if (Auth::user()->role == 1)
-        {
-            $request -> validate([
-                'hari_id'=> 'required',
-                'jam_id'=> 'required',
-                'ruangan_id'=> 'required',
-                'pengampu_id'=> 'required',
-                'reservasi_id'=> 'required'
+    {
+        if (Auth::user()->role == 1) {
+            $request->validate([
+                'hari_id' => 'required',
+                'jam_id' => 'required',
+                'ruangan_id' => 'required',
+                'pengampu_id' => 'required',
+                'reservasi_id' => 'required'
             ]);
             $add = Jadwal::create([
-            'hari_id'=>$request -> hari_id,
-            'jam_id' =>$request -> jam_id,
-            'ruangan_id'=> $request -> ruangan_id,
-            'pengampu_id'=> $request -> pengampu_id,
-            'reservasi_id'=> $request -> reservasi_id
+                'hari_id' => $request->hari_id,
+                'jam_id' => $request->jam_id,
+                'ruangan_id' => $request->ruangan_id,
+                'pengampu_id' => $request->pengampu_id,
+                'reservasi_id' => $request->reservasi_id
             ]);
             return $add;
         }
-        return['Anda tidak memiliki akses',];
-
+        return ['Anda tidak memiliki akses',];
     }
 
     /**
      * Display the specified resource.
      */
     public function show(string $id)
-    {   if (Auth::user()->role == 1)
-        {
-            $jadwalexist = Jadwal::where('id',$id)->first();
-            if ($jadwalexist){
+    {
+        if (Auth::user()->role == 1) {
+            $jadwalexist = Jadwal::where('id', $id)->first();
+            if ($jadwalexist) {
                 return new JadwalResource($jadwalexist);
             }
-            return['jadwal tidak ditemukan'];
+            return ['jadwal tidak ditemukan'];
         }
-        return['Anda tidak memiliki akses',];
+        return ['Anda tidak memiliki akses',];
     }
 
 
@@ -70,38 +69,37 @@ class JadwalController extends Controller
      * Update the specified resource in storage.
      */
     public function update(Request $request, string $id)
-    {   if (Auth::user()->role == 1)
-        {
-            $jadwalexist = Jadwal::where('id',$id)->first();
-            if ($jadwalexist)
-            {
-            $jadwalexist -> update([
-                'hari_id'       =>$request -> hari_id ?? $jadwalexist -> hari_id,
-                'jam_id'        =>$request -> jam_id ?? $jadwalexist ->jam_id,
-                'ruangan_id'    =>$request -> ruangan_id  ?? $jadwalexist -> ruangan_id,
-                'pengampu_id'   =>$request -> pengampu_id ?? $jadwalexist -> pengampu_id,
-                'reservasi_id'  =>$request -> reservasi_id ?? $jadwalexist -> reservasi_id
+    {
+        if (Auth::user()->role == 1) {
+            $jadwalexist = Jadwal::where('id', $id)->first();
+            if ($jadwalexist) {
+                $jadwalexist->update([
+                    'hari_id'       => $request->hari_id ?? $jadwalexist->hari_id,
+                    'jam_id'        => $request->jam_id ?? $jadwalexist->jam_id,
+                    'ruangan_id'    => $request->ruangan_id  ?? $jadwalexist->ruangan_id,
+                    'pengampu_id'   => $request->pengampu_id ?? $jadwalexist->pengampu_id,
+                    'reservasi_id'  => $request->reservasi_id ?? $jadwalexist->reservasi_id
                 ]);
-                return['jadwal berhasil di update',$jadwalexist,];
+                return ['jadwal berhasil di update', $jadwalexist,];
             }
-            return['Jadwal tidak ditemukan'];
+            return ['Jadwal tidak ditemukan'];
         }
-        return['Anda tidak memiliki akses'];
+        return ['Anda tidak memiliki akses'];
     }
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
-    {   if (Auth::user()->role == 1)
-        {
-            $jadwalexist = Jadwal::where('id','=',$id)->first();
-            if ($jadwalexist){
+    {
+        if (Auth::user()->role == 1) {
+            $jadwalexist = Jadwal::where('id', '=', $id)->first();
+            if ($jadwalexist) {
                 $jadwalexist->delete();
-                    return['Jadwal berhasil di hapus'];
+                return ['Jadwal berhasil di hapus'];
             }
             return ['jadwal tidak ditemukan'];
         }
-        return['Anda tidak memiliki akses'];
+        return ['Anda tidak memiliki akses'];
     }
 }
